+++ client/MemDB-Java-Client/src/main/java/es/memdb/messages/request/RequestNumberGenerator.java
@@ -1,11 +1,11 @@
 package es.memdb.messages.request;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public final class RequestNumberGenerator {
-    private final AtomicLong last = new AtomicLong(0);
+    private final AtomicInteger last = new AtomicInteger(0);
 
-    public long next() {
+    public int next() {
         return this.last.getAndIncrement();
     }
 }
